deploy/cpanel/requirements-cpanel.txt
# ═══════════════════════════════════════════════════════════════════════════════
# REQUIREMENTS PARA cPANEL - StrateKaz SGI
# ═══════════════════════════════════════════════════════════════════════════════
# Este archivo contiene las dependencias optimizadas para cPanel
# - Sin Redis (no disponible en hosting compartido)
# - PyMySQL en lugar de mysqlclient (no requiere compilación)
# - Sin herramientas de desarrollo
#
# Uso:
#   pip install -r requirements-cpanel.txt
# ═══════════════════════════════════════════════════════════════════════════════

# ═══════════════════════════════════════════════════
# CORE FRAMEWORK
# ═══════════════════════════════════════════════════
Django==5.0.9
djangorestframework==3.14.0

# ═══════════════════════════════════════════════════
# BASE DE DATOS
# ═══════════════════════════════════════════════════
# PyMySQL: Alternativa a mysqlclient que no requiere compilación
# mysqlclient requiere mysql-devel que no está en cPanel
PyMySQL==1.1.0

# ═══════════════════════════════════════════════════
# AUTENTICACIÓN Y SEGURIDAD
# ═══════════════════════════════════════════════════
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
cryptography==41.0.7
django-ratelimit==4.1.0
django-csp==3.8
bleach==6.1.0

# ═══════════════════════════════════════════════════
# CONFIGURACIÓN Y UTILIDADES
# ═══════════════════════════════════════════════════
python-decouple==3.8
python-dotenv==1.0.0
django-filter==23.5
django-auditlog==2.3.0
drf-spectacular==0.27.0

# ═══════════════════════════════════════════════════
# CELERY (modo EAGER - sin worker)
# ═══════════════════════════════════════════════════
# Incluimos Celery para compatibilidad de código
# pero funcionará en modo síncrono (CELERY_TASK_ALWAYS_EAGER=True)
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
# NO incluir: redis, flower (no funcionan en cPanel)

# ═══════════════════════════════════════════════════
# GENERACIÓN DE DOCUMENTOS
# ═══════════════════════════════════════════════════
# WeasyPrint puede requerir dependencias del sistema
# Si falla, comentar y usar alternativas
WeasyPrint==60.1
Pillow==10.1.0
qrcode[pil]==7.4.2
openpyxl==3.1.2

# ═══════════════════════════════════════════════════
# DEPLOYMENT
# ═══════════════════════════════════════════════════
# gunicorn no es necesario en cPanel (usa Passenger)
# pero lo incluimos por si se usa en otro contexto
whitenoise==6.6.0

# ═══════════════════════════════════════════════════
# MONITOREO (Opcional)
# ═══════════════════════════════════════════════════
sentry-sdk==1.40.0

# ═══════════════════════════════════════════════════
# NO INCLUIDOS EN cPANEL
# ═══════════════════════════════════════════════════
# redis - No hay servidor Redis en cPanel
# django-redis - Depende de Redis
# flower - Depende de Redis
# gunicorn - cPanel usa Passenger
# mysqlclient - Requiere compilación
# django-debug-toolbar - Solo desarrollo
# pytest-* - Solo testing
# black, ruff - Solo desarrollo
